HW6:almost done
